founder desktop fixed
