+++ src/lib/components/academy/founder.svelte
@@ -5,7 +5,7 @@
       "Our students go from zero coding experience to thriving tech careers. Join us and turn your tech dreams into reality!";
   </script>
 
-<section class="max-w-[360px] mx-auto text-center py-3">
+<section class="sm:hidden max-w-[360px] mx-auto text-center py-3">
   <!-- <p class="text-4xl font-500 font-[bai jamjuree] text-gray-900 mb-6 custom-underline text-center lg:text-left">
     Words from the <br>
     Founder and Mentor
@@ -16,6 +16,20 @@
     With our experience in <span class="!text-[#333333] !font-[500]"> building enterprise blockchain solutions</span>, we've witnessed its immense potential but also identified a significant skills gap <span class="!text-[#333333] !font-[500]">  demand for blockchain expertise </span> is rising, yet qualified talent remains scarce.  </p>
   <p class="text-[#5C5C5C] text-sm font-[300] font-rubik mb-[20px]"> <span class="!text-[#333333] !font-[500]">This program empowers the next generation of blockchain developers through a comprehensive curriculum </span> that goes beyond Bitcoin, delving into tokenomics, portfolio projects, and more</p>
   <p class="text-[#5C5C5C] text-sm font-[300] font-rubik mb-[20px]">
+    <span class="!text-[#333333] !font-[500]"> With hands-on projects and industry-relevant training, you'll gain future-proof skills </span> to thrive in this transformative field. Join us and take your career from ZERO to HERO!</p>
+</section>
+
+<section class="hidden sm:flex flex-col max-w-full h-auto mx-auto xl:h-[617px] text-left pt-8 py-3 mb-10 px-5 md:px-0">
+  <p class="text-5xl font-500 font-[bai jamjuree] text-[#333333] mb-6 custom-underline text-center lg:text-left">
+    Words from the 
+    Founder <br> and Mentor
+  </p>
+  <p class="text-[#5C5C5C] text-xl font-[300] font-rubik mb-[38px] mt-6">
+    At Timechain Labs, we proudly present <span class="!text-[#333333] !font-[500]"> The Blockchain Mastery Program</span>, designed to accelerate your career in the evolving tech industry.  </p>
+  <p class="text-[#5C5C5C] text-xl font-[300] font-rubik mb-[38px]">
+    With our experience in <span class="!text-[#333333] !font-[500]"> building enterprise blockchain solutions</span>, we've witnessed its immense potential but also identified a significant skills gap <span class="!text-[#333333] !font-[500]">  demand for blockchain expertise </span> is rising, yet qualified talent remains scarce.  </p>
+  <p class="text-[#5C5C5C] text-xl font-[300] font-rubik mb-[38px]"> <span class="!text-[#333333] !font-[500]">This program empowers the next generation of blockchain developers through a comprehensive curriculum </span> that goes beyond Bitcoin, delving into tokenomics, portfolio projects, and more</p>
+  <p class="text-[#5C5C5C] text-xl font-[300] font-rubik mb-[38px]">
     <span class="!text-[#333333] !font-[500]"> With hands-on projects and industry-relevant training, you'll gain future-proof skills </span> to thrive in this transformative field. Join us and take your career from ZERO to HERO!</p>
 </section>
 
@@ -36,10 +50,10 @@
       background-color:#059E93;
       opacity:0.7;    
       }
-      /* .custom-underline {
+      .custom-underline {
         text-decoration: underline;
         text-decoration-color: #059E93;
         text-decoration-thickness: 5px;
         text-underline-offset: 1.2px; 
-      } */
+      }
 </style>
